Replace hardcoded option numbers by constants from `OptionsProto`

+++ gradle-plugins/model-compiler/src/main/java/io/spine/gradle/compiler/annotation/Annotator.java
@@ -171,7 +171,7 @@ abstract class Annotator<O extends ExtendableMessage, D extends GeneratedMessage
      *
      * @return the option number
      */
-    protected final long getOptionNumber() {
+    protected final int getOptionNumber() {
         return option.getNumber();
     }
